Fix CDKGSessionHandler thread names (#3896)
Some compilers seems to have an issue with this implicit conversion producing garbage instead of numbers.

+++ src/llmq/quorums_dkgsessionhandler.cpp
@@ -142,7 +142,7 @@ void CDKGSessionHandler::StartThread()
         throw std::runtime_error("Tried to start an already started CDKGSessionHandler thread.");
     }
 
-    std::string threadName = strprintf("q-phase-%d", params.type);
+    std::string threadName = strprintf("q-phase-%d", (uint8_t)params.type);
     phaseHandlerThread = std::thread(&TraceThread<std::function<void()> >, threadName, std::function<void()>(std::bind(&CDKGSessionHandler::PhaseHandlerThread, this)));
 }
